feat: rename secrets module to secret
