+++ packages/core/src/secret/SecretManager.test.ts
